Turn registry reader into a plugin
This will require some configuration in the near future so that ie you
could restrict access only to explicitly specified relations:

  # would only expose :users and :tasks relations
  use :relation_registry, [:users, :tasks]

Getting rid of method_missing would also be nice
